Refactor dashboard page into components
