+++ flask_react/app/dashboard/sample-data.ts
@@ -1,0 +1,323 @@
+import type { Document, DocumentUpdate, Transaction, AuditEvent } from "./types"
+
+export const allDocuments: Document[] = [
+  {
+    id: 1,
+    title: 'Market Research Q4',
+    description: 'Detailed analysis of Q4 market trends',
+    docType: 'Research',
+    region: 'Global',
+    status: 'active',
+    price: 500,
+    duration: '1 year',
+    scope: 'Comprehensive',
+    trending: true,
+    relationship: 'owned',
+    isLicensed: false,
+    isShared: true,
+    expirationDate: '2025-12-31',
+    sharedWith: [
+      { name: 'Meridian Capital', type: 'Asset Management Firm', since: '2 days ago', expires: '2025-12-31' },
+      { name: 'CBRE', type: 'Real Estate Brokers', since: '1 week ago', expires: '2025-12-31' },
+    ],
+    licensedTo: [],
+    owner: 'You',
+    revenue: 0,
+  },
+  {
+    id: 2,
+    title: 'Industrial Property Valuations',
+    description: 'Valuations of industrial properties',
+    docType: 'Valuation',
+    region: 'North America',
+    status: 'active',
+    price: 1000,
+    duration: '2 years',
+    scope: 'Regional',
+    trending: false,
+    relationship: 'owned',
+    isLicensed: true,
+    isShared: true,
+    expirationDate: '2026-07-15',
+    sharedWith: [{ name: 'MRI Software', type: 'Lease Administrator', since: '3 days ago', expires: '2026-07-15' }],
+    licensedTo: [
+      { name: 'Brookfield Asset Management', type: 'Asset Management Firm', since: '2 hours ago', expires: '2026-07-15', amount: '$1000 USDC' },
+    ],
+    owner: 'You',
+    revenue: 1000,
+  },
+  {
+    id: 3,
+    title: 'Premium Office Lease Comps',
+    description: '20,000 sqft office space lease data with 10-year terms',
+    docType: 'Lease',
+    region: 'Northeast',
+    status: 'active',
+    price: 200,
+    duration: '12 months',
+    scope: 'Internal analytics',
+    trending: true,
+    relationship: 'personal-licensed',
+    isLicensed: true,
+    isShared: false,
+    expirationDate: '2025-07-10',
+    sharedWith: [],
+    licensedTo: [],
+    owner: '0xBrokerWallet',
+    revenue: 0,
+    licenseAmount: '$200 USDC',
+  },
+  {
+    id: 4,
+    title: 'Retail Space Analysis',
+    description: 'Comprehensive retail market analysis for downtown area',
+    docType: 'Analysis',
+    region: 'Midwest',
+    status: 'active',
+    price: 0,
+    duration: '6 months',
+    scope: 'Market research',
+    trending: false,
+    relationship: 'shared',
+    isLicensed: false,
+    isShared: true,
+    expirationDate: '2025-06-15',
+    sharedWith: [],
+    licensedTo: [],
+    owner: 'CoStar Group',
+    revenue: 0,
+  },
+  {
+    id: 5,
+    title: 'Downtown Office Market Report',
+    description: 'Quarterly analysis of downtown office market conditions',
+    docType: 'Report',
+    region: 'West Coast',
+    status: 'active',
+    price: 750,
+    duration: '18 months',
+    scope: 'Regional analysis',
+    trending: false,
+    relationship: 'firm-licensed',
+    isLicensed: true,
+    isShared: false,
+    expirationDate: '2026-01-15',
+    sharedWith: [],
+    licensedTo: [],
+    owner: 'JLL Research',
+    revenue: 0,
+    licenseAmount: '$750 USDC',
+    firmAccess: true,
+  },
+  {
+    id: 6,
+    title: 'Internal Property Assessment',
+    description: 'Confidential assessment for internal use only',
+    docType: 'Assessment',
+    region: 'Local',
+    status: 'active',
+    price: 0,
+    duration: 'Permanent',
+    scope: 'Internal use',
+    trending: false,
+    relationship: 'owned',
+    isLicensed: false,
+    isShared: false,
+    expirationDate: 'N/A',
+    sharedWith: [],
+    licensedTo: [],
+    owner: 'You',
+    revenue: 0,
+  },
+  {
+    id: 7,
+    title: 'Draft Market Analysis',
+    description: 'Work-in-progress market analysis document',
+    docType: 'Draft',
+    region: 'Regional',
+    status: 'draft',
+    price: 0,
+    duration: 'N/A',
+    scope: 'Internal review',
+    trending: false,
+    relationship: 'owned',
+    isLicensed: false,
+    isShared: false,
+    expirationDate: 'N/A',
+    sharedWith: [],
+    licensedTo: [],
+    owner: 'You',
+    revenue: 0,
+  },
+  {
+    id: 8,
+    title: 'Firm Portfolio Analysis',
+    description: "Comprehensive analysis of firm's real estate portfolio",
+    docType: 'Portfolio',
+    region: 'National',
+    status: 'active',
+    price: 0,
+    duration: 'Permanent',
+    scope: 'Firm-wide access',
+    trending: false,
+    relationship: 'firm-owned',
+    isLicensed: false,
+    isShared: false,
+    expirationDate: 'N/A',
+    sharedWith: [],
+    licensedTo: [],
+    owner: 'Your Firm',
+    revenue: 0,
+  },
+  {
+    id: 9,
+    title: 'Regional Market Intelligence',
+    description: 'Firm-owned market intelligence database',
+    docType: 'Database',
+    region: 'Multi-region',
+    status: 'active',
+    price: 0,
+    duration: 'Permanent',
+    scope: 'Firm-wide access',
+    trending: false,
+    relationship: 'firm-owned',
+    isLicensed: true,
+    isShared: true,
+    expirationDate: 'N/A',
+    sharedWith: [{ name: 'Partner Firms', type: 'Strategic Partners', since: '1 month ago', expires: '2026-12-31' }],
+    licensedTo: [
+      { name: 'External Consultants', type: 'Consulting Firm', since: '2 weeks ago', expires: '2025-12-31', amount: '$2500 USDC' },
+    ],
+    owner: 'Your Firm',
+    revenue: 2500,
+  },
+]
+
+export const documentUpdates: DocumentUpdate[] = [
+  {
+    id: 'lease_123',
+    title: 'Premium Office Lease Comps',
+    lastActivity: {
+      action: 'RELEASE_ESCROW',
+      stage: 'Marketplace & licensing',
+      timestamp: '2 hours ago',
+      color: 'bg-green-500',
+    },
+    totalEvents: 8,
+    hasMoreEvents: true,
+  },
+  {
+    id: 'market_456',
+    title: 'Q4 Market Analysis Report',
+    lastActivity: {
+      action: 'DATA_ACCESS_LOG',
+      stage: 'Data access monitoring',
+      timestamp: '30 minutes ago',
+      color: 'bg-orange-500',
+    },
+    totalEvents: 15,
+    hasMoreEvents: true,
+  },
+  {
+    id: 'appraisal_789',
+    title: 'Industrial Property Valuations',
+    lastActivity: {
+      action: 'ABSTRACT_VALIDATE',
+      stage: 'Structured data capture',
+      timestamp: '1 day ago',
+      color: 'bg-purple-500',
+    },
+    totalEvents: 4,
+    hasMoreEvents: true,
+  },
+  {
+    id: 'office_101',
+    title: 'Office Building Analysis',
+    lastActivity: {
+      action: 'REGISTER_ASSET',
+      stage: 'Document origination',
+      timestamp: '5 hours ago',
+      color: 'bg-blue-500',
+    },
+    totalEvents: 2,
+    hasMoreEvents: false,
+  },
+]
+
+export const marketplaceTransactions: Transaction[] = [
+  {
+    id: 1,
+    type: 'credit',
+    description: 'License fee received',
+    asset: 'Premium Office Lease Comps',
+    amount: '$200 USDC',
+    timestamp: '2 hours ago',
+    status: 'completed',
+    counterparty: 'Blackstone Real Estate',
+  },
+  {
+    id: 2,
+    type: 'debit',
+    description: 'License purchase',
+    asset: 'Q4 Market Analysis Report',
+    amount: '$500 USDC',
+    timestamp: '5 hours ago',
+    status: 'completed',
+    counterparty: 'CoStar Group',
+  },
+  {
+    id: 3,
+    type: 'credit',
+    description: 'Escrow release',
+    asset: 'Industrial Property Valuations',
+    amount: '$350 USDC',
+    timestamp: '1 day ago',
+    status: 'completed',
+    counterparty: 'Brookfield Asset Management',
+  },
+  {
+    id: 4,
+    type: 'debit',
+    description: 'Platform fee',
+    asset: 'Office Building Analysis',
+    amount: '$25 USDC',
+    timestamp: '2 days ago',
+    status: 'completed',
+    counterparty: 'Atlas DAO Treasury',
+  },
+]
+
+export const auditTrail: AuditEvent[] = [
+  {
+    stage: 'Document origination',
+    action: 'REGISTER_ASSET',
+    timestamp: '2025-07-10T18:22:05Z',
+    actor: '0xBrokerWallet',
+    txHash: '0xabc123...',
+    status: 'completed',
+  },
+  {
+    stage: 'Structured data capture',
+    action: 'AI_ABSTRACT_SUBMIT',
+    timestamp: '2025-07-10T18:22:55Z',
+    actor: 'AtlasAIService',
+    txHash: '0x789aaa...',
+    status: 'completed',
+  },
+  {
+    stage: 'Marketplace & licensing',
+    action: 'CREATE_LICENSE_OFFER',
+    timestamp: '2025-07-11T02:15:00Z',
+    actor: '0xBrokerWallet',
+    txHash: '0x112233...',
+    status: 'completed',
+  },
+  {
+    stage: 'Payment & settlement',
+    action: 'RELEASE_ESCROW',
+    timestamp: '2025-07-11T05:03:05Z',
+    actor: 'EscrowContract',
+    txHash: '0x667788...',
+    status: 'completed',
+  },
+]
